Embassy Target Server Impl (#9)
This PR adds an initial Target Server impl, using `embassy-usb` as a baseline, and providing a routing macro to
help reduce boilerplate.

This is NOT release-ready, as we still have patched dependencies, and we are waiting on some upstream PRs to
`embassy-usb` and `static-cell` to settle. I'll address these issues in follow-up PRs before a `0.4.3` or `0.5` release.