`define_dispatch`: add docs for app `type`
since the `type` generated for the app is marked as `pub` this triggers
the `missing_docs` lint (if enabled) and thus prevents
`#![deny(missing_docs)]` from being used in any firmware which uses
`postcard-rpc`.
by generating this doc comment we now make the lint happy.

the alternative would have been to add `#[doc(hidden)]`, however there's
no reason to hide it and it has been present in the docs so far.

note that macro variables are not replaced in normal doc comments
(written using `///`), however these anyway get converted to
`#[doc="..."]`, thus we can make use of that. the idea for that came
from SO: https://stackoverflow.com/a/78426606

the `minimal` example has been modified to include
`#[deny(missing_docs)]` so as to ensure that no future modification will
break this again.

resolves #106