add `postcard-schema` re-export for macros
the macros reference `postcard-schema` and expect their callers to add
the dependency.
this has two downsides:
* this dependency leaks to the caller
* we have no guarantee as to which version the caller adds (and with
  which features), thus we don't know whether it'll actually compile

by re-exporting it and referencing our re-export we can resolve this
issue. the re-export is hidden from the docs as consumers are not
supposed to use it directly.

note that the ICD example has not been modified since it explicitly uses
`#[derive(Schema)]` and thus directly depends on the crate.

resolves #105