fix build warning about unused import on windows (#122)
* fix build warning about unused import on windows

`nusb::InterfaceInfo` was always `use`ed, however it was only referenced
on non-windows OSes, thus causing an "unused import" build warning when
building on windows.

* deny unused imports

with the previous commit an unused import warning was fixed on windows.
to prevent this from happening again a new deny rule is added.

important: to ensure that this really works the CI needs to check all
config options, same as for other compile errors.