add missing `impl Error`s (#130)
use `thiserror` to do this as it's already used in the crate. this
updates it from v1 to v2 and also uses it in `no_std` environments as
that's now supported by `thiserror`.

this was missing and thus it wasn't possible to use e.g. `HostErr` in a
`#[from]` with `thiserror`.